+++ pybricksdev/ble/lwp3/__init__.py
@@ -1,11 +1,13 @@
 # SPDX-License-Identifier: MIT
-# Copyright (c) 2021 The Pybricks Authors
+# Copyright (c) 2021-2022 The Pybricks Authors
 
 """This module and its submodules are used for Bluetooth Low Energy
 communications with devices that provide the LEGO Wireless Protocol v3.
 """
 
 from enum import IntEnum
+
+from .bytecodes import Capabilities, HubKind, LastNetwork, Status, Version
 
 # LEGO Wireless Protocol v3 is defined at:
 # https://lego.github.io/lego-ble-wireless-protocol-docs/
@@ -105,3 +107,91 @@ class BootloaderError(IntEnum):
 
     UNKNOWN_COMMAND = 0x05
     """The command was not recognized."""
+
+
+class BootloaderAdvertisementData:
+    """
+    The 6-byte manufacturer-specific advertisement data sent when a hub in
+    bootloader mode is advertising.
+    """
+
+    def __init__(self, data: bytes) -> None:
+        if len(data) != 6:
+            raise ValueError("expecting 6 bytes of data")
+
+        self._data = data
+
+    def __bytes__(self) -> bytes:
+        return self._data
+
+    @property
+    def version(self) -> Version:
+        """
+        Gets the bootloader software version.
+        """
+        return Version(int.from_bytes(self._data[:4], "little"))
+
+    @property
+    def hub_kind(self) -> HubKind:
+        """
+        Gets the hub type identifier.
+        """
+        return HubKind(self._data[4])
+
+    @property
+    def hub_capabilities(self) -> Capabilities:
+        """
+        Gets the hub capability flags.
+        """
+        return Capabilities(self._data[5])
+
+
+class AdvertisementData:
+    """
+    The 6-byte manufacturer-specific advertisement data sent when a hub running
+    official LEGO firmware is advertising.
+    """
+
+    def __init__(self, data: bytes) -> None:
+        if len(data) != 6:
+            raise ValueError("expecting 6 bytes of data")
+
+        self._data = data
+
+    def __bytes__(self) -> bytes:
+        return self._data
+
+    @property
+    def is_button_pressed(self) -> bool:
+        """
+        Gets the state of the button on the hub.
+        """
+        return bool(self._data[0])
+
+    @property
+    def hub_kind(self) -> HubKind:
+        """
+        Gets the hub type identifier.
+        """
+        return HubKind(self._data[1])
+
+    @property
+    def hub_capabilities(self) -> Capabilities:
+        """
+        Gets the hub capability flags.
+        """
+        return Capabilities(self._data[2])
+
+    @property
+    def last_network(self) -> LastNetwork:
+        """
+        Gets the ID of the last network the hub was connected to.
+        """
+        return LastNetwork(self._data[3])
+
+    @property
+    def status(self) -> Status:
+        """
+        Gets the status of the advertising hub.
+        """
+        return Status(self._data[4])