+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added ``ble.lpw3.AdvertisementData`` class.
+- Added ``ble.lpw3.BootloaderAdvertisementData`` class.
+
 ## [1.0.0-alpha.21] - 2022-01-12
 
 ### Changed